Trying to fix cloudformation stack
